+++ croud/cmd.py
@@ -397,56 +397,9 @@ def consumer_id_arg(req_args: _ArgumentGroup, opt_args: _ArgumentGroup) -> None:
     )
 
 
-def consumer_eventhub_dsn_arg(
-    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool = True
-) -> None:
-    group = req_args if required else opt_args
-    group.add_argument(
-        "--consumer-eventhub-dsn",
-        type=str,
-        help="The connection string to the Azure EventHub from which to consume.",
-        required=required,
-    )
-
-
-def consumer_eventhub_consumer_group_arg(
-    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool = True
-) -> None:
-    group = req_args if required else opt_args
-    group.add_argument(
-        "--consumer-eventhub-consumer-group",
-        type=str,
-        help="The consumer group for the Azure EventHub from which to consume.",
-        required=required,
-    )
-
-
-def consumer_eventhub_lease_storage_dsn_arg(
-    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool = True
-) -> None:
-    group = req_args if required else opt_args
-    group.add_argument(
-        "--consumer-eventhub-lease-storage-dsn",
-        type=str,
-        help=(
-            "The connection string to an Azure storage account to use as lease storage."
-        ),
-        required=required,
-    )
-
-
-def consumer_eventhub_lease_storage_container_arg(
-    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool = True
-) -> None:
-    group = req_args if required else opt_args
-    group.add_argument(
-        "--consumer-eventhub-lease-storage-container",
-        type=str,
-        help=(
-            "The container name in the lease storage for the Azure EventHub "
-            "consumer to use."
-        ),
-        required=required,
+def consumer_name_arg(req_args: _ArgumentGroup, opt_args: _ArgumentGroup) -> None:
+    req_args.add_argument(
+        "--consumer-name", type=str, help="The consumer name to use.", required=True
     )
 
 
@@ -471,6 +424,69 @@ def consumer_table_arg(
         type=str,
         help="The CrateDB database table used by the Azure EventHub consumer.",
         required=required,
+    )
+
+
+def eventhub_dsn_arg(
+    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool = True
+) -> None:
+    group = req_args if required else opt_args
+    group.add_argument(
+        "--eventhub-dsn",
+        type=str,
+        help="The connection string to the Azure EventHub from which to consume.",
+        required=required,
+    )
+
+
+def eventhub_consumer_group_arg(
+    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool = True
+) -> None:
+    group = req_args if required else opt_args
+    group.add_argument(
+        "--eventhub-consumer-group",
+        type=str,
+        help="The consumer group of the Azure EventHub from which to consume.",
+        required=required,
+    )
+
+
+def lease_storage_dsn_arg(
+    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool = True
+) -> None:
+    group = req_args if required else opt_args
+    group.add_argument(
+        "--lease-storage-dsn",
+        type=str,
+        help=(
+            "The connection string to an Azure storage account to use as lease storage."
+        ),
+        required=required,
+    )
+
+
+def lease_storage_container_arg(
+    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool = True
+) -> None:
+    group = req_args if required else opt_args
+    group.add_argument(
+        "--lease-storage-container",
+        type=str,
+        help=(
+            "The container name in the lease storage for the Azure EventHub "
+            "consumer to use."
+        ),
+        required=required,
+    )
+
+
+def num_instances_arg(req_args: _ArgumentGroup, opt_args: _ArgumentGroup) -> None:
+    opt_args.add_argument(
+        "--num-instances",
+        type=int,
+        help="The number of instances to deploy.",
+        default=1,
+        required=False,
     )